pyproject.toml
[project]
name = "paperflow"
version = "0.1.0"
description = "Academic journal aggregation CLI with bilingual translation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.9.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "feedparser>=6.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "sqlalchemy>=2.0.0",
    "ollama>=0.1.0",
]

[project.scripts]
paperflow = "paperflow.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/paperflow"]

[tool.hatch.build.targets.sdist]
include = ["src/paperflow"]
